Added AJAX compatibility and security to binaries
Execute this SQL to update the module folder name in the database!
UPDATE `zpanel_core`.`x_modules` SET `mo_folder_vc` = 'zpanel_config'
WHERE `x_modules`.`mo_folder_vc` = 'zpanelconfig';

Further notes:
- zpanel_config module updated to use the implemented secure way of
running the daemon.
- files in the bin folder cannot be accessed via apache (thus these
can't be executed anymore via control.example.com/bin/daemon.php)
- Tested on Ubuntu and should work on CentOS. Windows needs testing, the
only thing that could go wrong with Windows is that the daemon file may
not be called in Windows if the default directory in php is not the
panel one.
