+++ inc/init.inc.php
@@ -140,5 +140,4 @@ if (!isset($_SESSION['zpuid'])) {
 
 runtime_hook::Execute('OnBeforeControllerInit');
 $controller->Init();
-ui_templateparser::Generate("etc/styles/" . ui_template::GetUserTemplate());
 ?>
