+++ modules/zpanel_config/assets/ajaxsbmt.js
