+++ modules/zpanel_config/assets/icon.png
