--- modules/zpanelconfig/module.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<module>
-    <name>ZPanel Config</name>
-    <version>100</version>
-    <defaultcat>Server Admin</defaultcat>
-    <desc>Changes made here affect the entire ZPanel configuration, please double check everything before saving changes.</desc>
-    <authorname>ZPanel Project</authorname>
-    <authoremail>developers@zpanelcp.com</authoremail>
-    <authorurl>http://www.zpanelcp.com/</authorurl>
-    <updateurl>http://www.zpanelcp.com/uds/core.xml</updateurl>
-</module>
